Kotlin Lab1

+++ src/Lab1/pr1_3.kt
@@ -3,6 +3,6 @@ package Lab1
 import kotlin.math.exp
 
 fun main() {
-    val num = String.format("%.3f", exp(1.0)).toDouble()
+    val num = String.format("%.3f", exp(1.0))
     println("Your number is $num")
 }
